>From juliane@dcd.wa.gov.au Mon Sep 20 20:36:49 1993 Date: Tue, 21 Sep 93 10:25:12 WST As I mentionned yesterday, I have made an interim release of the new scsi code.
This is basically a half-way point in the rewrite. It impliments
more dynamic allocation and some new structures. The basic logic
of how the system fits together has not changed.

Once this set of changes has been verified as having "NOT BROKEN" the system,
then I will move on to other changes that first required these changes.

BIG NOTE:  THIS IS COMMITED ONTO A BRANCH FOR TESTING ONLY!