Changes in order to compile GCC 3.0.2.
Really irritating changes are the "forced" layering of malloc + friends
in order to use the GNU versions.  Sorry, we have a *very* fine malloc,
and we will use it.  Period.  Even more irritating is that the GNU people
now want to replace ctype also!!  So we partially dike it out here.